refactor: 패키지 변경, 인자 타입 통일
